feat: implement ibc core paths (#274)
* chore: update storage for block time

* feat: implement storage modifiers for expected time per block

* refactor: cargo clippy

* chore: update codebase and remove uncessary import and refrencing

* test: add unit test

* fix: typos

* fix: typos

+++ contracts/cosmwasm-vm/cw-ibc-core/tests/channel/test_timeout.rs
@@ -335,7 +335,7 @@ fn test_timeout_packet_validate_to_light_client() {
     contract
         .ibc_store()
         .expected_time_per_block()
-        .save(deps.as_mut().storage, &(env.block.time.seconds() as u128))
+        .save(deps.as_mut().storage, &(env.block.time.seconds()))
         .unwrap();
 
     let res = contract.timeout_packet_validate_to_light_client(deps.as_mut(), info, &msg);
